orders/requirements.txt
astroid==2.5
attrs==20.3.0
autopep8==1.5.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
fastapi==0.63.0
flake8==3.8.4
h11==0.12.0
idna==2.10
iniconfig==1.1.1
isort==5.7.0
lazy-object-proxy==1.5.2
mccabe==0.6.1
motor==2.3.1
packaging==20.9
pep8==1.7.1
pluggy==0.13.1
py==1.10.0
pycodestyle==2.6.0
pydantic==1.7.3
pyflakes==2.2.0
PyJWT==2.0.1
pylint==2.7.0
pymongo==3.11.3
pyparsing==2.4.7
pytest==6.2.2
requests==2.25.1
rope==0.18.0
starlette==0.13.6
toml==0.10.2
urllib3==1.26.3
uvicorn==0.13.4
wrapt==1.12.1
